+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # NEXT RELEASE
+
+# v0.31.0
 * Increase default `TIME_DECAY` from 0.05 to 0.15
 * Change constructor argument from `language` to `locale`
 * Add `Account.noteWithAccountInfo()` method